Name compiled function with the jade template's filename
The jade --client function compiles a template into a javascript
function for client-side injection. While the generated javascript
file is named from the jade template, the javascript function within
that file is always named 'template'. This is a problem if we want
to combine a number of these functions in one file.

The proposal is to name the function with the same basename as
the file which contains it. For example, the partial jade template
in file 'user.jade' would compile to js file 'user.js' with the
function 'userTemplate'. This would allow concatenation in our
grunt/gulp workflow without duplicated function names.